pass versionId data between version and configure version

+++ app/src/main/java/com/waelkhelil/sayara_dz/database/model/Version.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "version")
 data class Version(
-    @PrimaryKey var id:Long,
+    @PrimaryKey var id:Int,
     @ColumnInfo var name:String,
     @ColumnInfo var url: String,
     @ColumnInfo var price: String,
